refactor: reorganize include structure and create fps_types.h
Phase 1 - Leaf header cleanup:
- milkDebugTools.h: remove unused string.h
- milkdata.h: remove 6 redundant system headers
- processtools.h: remove 4 redundant system headers
- CLIcore.h/CLIcore_standalone.h: remove 7 redundant system
  headers, keep string.h (used by INIT_MODULE_LIB macro)
- Add explicit string.h to 4 .c files that relied on
  transitive includes

Phase 2 - Split fps.h:
- Create fps_types.h with FPS type definitions, structs,
  constants, and flags (no API or macros)
- Refactor fps.h to delegate types to fps_types.h
- Add fps_types.h to CMake install list

Code that only needs FPS types can now include fps_types.h
instead of the full fps.h, avoiding 35+ sub-header includes.
