+++ src/cli/CLIcore/CLIcore.h
@@ -23,17 +23,10 @@
 #else /* full CLI mode */
 
 
-#include <errno.h>
-#include <sched.h>
-#include <semaphore.h>
-#include <signal.h>
 #include <stdint.h>
 #include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include <sys/types.h>
-#include <unistd.h>
 
 #include "config.h"
