+++ src/cli/CLIcore/CLIcore_standalone.h
@@ -17,17 +17,10 @@
 #ifndef CLICORE_STANDALONE_H
 #define CLICORE_STANDALONE_H
 
-#include <errno.h>
-#include <sched.h>
-#include <semaphore.h>
-#include <signal.h>
 #include <stdint.h>
 #include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include <sys/types.h>
-#include <unistd.h>
 
 #include "config.h"
