+++ src/coremods/COREMOD_memory/create_variable.c
@@ -3,6 +3,8 @@
  * @brief   create variables
  */
 
+#include <string.h>
+
 #ifdef MILK_NO_CLI
 #include "CLIcore_standalone.h"
 #else
