+++ src/coremods/COREMOD_memory/variable_ID.c
@@ -3,6 +3,8 @@
  * @brief   find variable ID(s) from name
  */
 
+#include <string.h>
+
 #ifdef MILK_NO_CLI
 #include "CLIcore_standalone.h"
 #else
