+++ src/engine/libmilkdata/milkDebugTools.h
@@ -9,7 +9,6 @@
 #define _MILKDEBUGTOOLS_H
 
 #include <errno.h>
-#include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
